v0.2.0 - higlighting of the cells.
Getting the indeces of the cubes of the figure is added. It allows
implementing the highlighting of the projections of the cubes on the
corresponding walls.

Further actions:
1) amend O, L figures, such that they are rotated around one of their
vertices;

2) check that the figure doesn't go beyond boundaries when moved;

3) check that the figure doesn't go beyond the scene boundaries when
rotated. If it does, then either prevent from rotating or bring it back
after the rotation has occurred;

4) collision detection: if the cell under the cubes cell is filled or is
outside the vertical range, then fix the figure, make sound, add cubes to
the fixed cubes, create a new figure;

5) draw bottom plane highlighting on the topmost planes of the fixed
cubes;

6) layer annihilation: if the plane is filled, then delete it together
with its cubes.